Add EnsureHasNoProfileAnnotation.
This is a meta-annotation, used to annotate new annotations of the form
EnsureHasNoXProfile. This will make it easy to create new annotations of
that form without needing to modify Harrier.

This is the pattern we should use for custom profile types - rather than
a generic annotation with an argument - because the new annotation types
can be better targeted by host-side configuration.

Test: atest HarrierTest
Bug: 183386080
Change-Id: I71ba5e78890fa644d6462b9d18ae44ad936dc3af