Some ASurfaceControlTest cleanup
1. Remove the animationFactory passed to verifyTest. This isn't needed
for this set of tests and just slows down the tests significantly.
Instead, this set of tests will just test that a single frame showed the
correct content.

2. Use DEFAULT_LAYOUT_WIDTH and DEFAULT_LAYOUT_HEIGHT as the buffer size
sent to SurfaceControlTestCase. This controls the size of the
SurfaceView. Previously, the buffer size was larger and then getting
scaled down, which was causing confusion in the tests. This simplies the
tests since now the expected size is much clearer

Test: ASurfaceControlTest
Fixes: 184074163
Change-Id: I8063f5bbb2474618162c0eed82800bf2efbf3e3c