Fixed testSetMeteredDataDisabledPackages for automotive.
This test uses WifiManager to get and set info about the active
wifi conenction. The 'setters' must always be called by the device
owner, while the 'getter' must be called by profile owner on headless
system user mode (as on automotive the system user has location
disabled, which would refact the SSID and cause the test to fail).

Test: a CtsDevicePolicyManagerTestCases:com.android.cts.devicepolicy.MixedDeviceOwnerTest#testSetMeteredDataDisabledPackages # on phone and automotive
Fixes: 132373384

Change-Id: I9a2f4e9e69be1e35367ac64fb74bcc0d4bac02e0