Revert "Merge "Set temporary target SDK until S SDK finalised." into sc-dev am: 7572eb0aa5"
This reverts commit 7ff96253355ba436321e5d90febcf7b2543c86c3.

Reason for revert: This change was needed so our tests from S branches can be installed and run on R devices (b/182960918). This was a temporary workaround until S SDK is finalised (b/187502983). Reverting in master since this requirement does not apply in master and tagetting "current" is what we want to do in master.

Change-Id: I38aa6c3e4b205fed110729649a7903415c695380