Fix flaky testTranslationCapabilityUpdateListener.
This test tries to register DeviceTranslationCapabilityUpdateListener
then verifies the listener is called and unregister then verifies the
listener doesn't be called. There is a very small possibility that
the unregister doesn't finish yet, then test let TranslationService
to update translation capabilities.

To fix the issue, we let the test to wait a timeout then starts to
verify it.

Bug: 187623816
Test: atest CtsTranslationTestCases:TranslationManagerTest#\
testTranslationCapabilityUpdateListener --iterations 100

Change-Id: I4132389838397e408db020c447407d504af0cbc3