Fixed SecurityLoggingTest for automotive.
On headless system user mode, the test runs on current user (which
has a PO), the device owner is set on system user, and DpmWrapper
tunnels calls from former to the latter. Hence, this test was failing
because the events were generated by DO, but the test was expecting
them to be generated by the PO.

This CL also improves the logging statements in case of errors.

Test: atest com.android.cts.devicepolicy.MixedDeviceOwnerTest#testSecurityLoggingWithSingleUser # on phone and automotive
Test: atest com.android.cts.devicepolicy.OrgOwnedProfileOwnerTest#testSecurityLogging,testSecurityLoggingDelegate # on phone

Fixes: 187984747

Change-Id: Id79c5770cdc34689dcaefaed9b8a8fb2f2d2f9ce