Verify isUiContext in InputMethodServiceTest
After migrating InputMethodService to WindowProviderService, the API
isUiContext is implemented by its base context vs hardcoded to true.
Therefore, the test testIsUiContextOnInputMethodService, which only
verify the InputMethodService ctr, will fail
because we don't attach the base context and it's not possible to
obtain the real InputMethodService instance in test.
Switch to verify isUiContext on InputMethodServiceTest.

Test: atest ContextIsUiContextTest
Test: atest InputMethodServiceTest
Bug: 159767464
Change-Id: I4adc955329eab7a00519d22cc859e98a3e402892