Policy transparency tests: Safeguard against settings intent not handled
Policy transparency tests offer a button to open the corresponding
settings page. The used Settings.ACTION_*_SETTINGS intents are generally
optional and may not be implemented on all devices, so CtsVerifier
should safeguard against the intent not being handled.
If starting the specific activity fails, fall back to opening the main
settings activity.

Bug: 181201913
Test: CtsVerifier
Change-Id: I4b705e7808b8a4d87392649f9c531d3ab784bc78