Filter out tasks not in default display
IntentTests work by use hardcoded before/after states in configuration
files. This makes the assumption that there won't be tasks running in
other displays which is not the case for automotive. In automotive, we
may have a cluster activity running in the cluster display.

To address this, we add a filter for the default display before
comparing the expected state.

Bug: 184786379
Test: atest IntentTests
Change-Id: I406775458e3c892db514ae6ec95739b8a48c9650