Wait to call @hide methods in ACSD test
In b/178989054 some tests in CloseSystemDialogsTest are flaky and
resulting in NoSuchMethodErrors. I believe this can be due to a race
between we enabling those methods in ART and we calling them before that
action completes. So, waiting up to 3s after calling the service until
we are sure we can call @hide methods (ag/13327152 explains why we need
to call those methods.

Test: atest CtsAppTestCases:android.app.cts.CloseSystemDialogsTest
Fixes: 178989054
Change-Id: Ica048204a34ced67f83961d409a1afd4074327af