Fix testValueTracksState
For some devices, it may take a little more time for statusbar  expand-settings or collapse.
so we need add some delay Wait for the animation to end. If the animation does not end,
it will not trigger onStartListening in TileService and causes the test to fail.

Bug:191184365

Test: run cts -m CtsAppTestCases -t android.app.cts.BooleanTileServiceTest#testValueTracksState

Signed-off-by: duhongfeng <duhongfeng@xiaomi.com>
Change-Id: Ib110eefd1790bc8be1af7a703bbdb881b6f0b3e8
(cherry picked from commit ca19b28ead71275c6b247516ed000ccddf661cf0)