Fix ActivityTransitionTests failure on slow devices
ActivityTransitionTests#taskTransitOverrideDisabled and
ActivityTransitionTests#taskTransitOverride check which animation runs
by measuring actual duration which the device spends to finish the
animation.

On slow devices, it can take longer to run those animations. So we made
the duraiton of one of the animation longer, and relaxed the threashold
for the check.

Bug: 184780552
Test: atest ActivityTransitionTests
Change-Id: Ifc221b8e49e197effb102fedc1cb8830c408a38e