Fix AppOps atom tests for hotword appop transform
Only the current HotwordDetectionService uid can note
OP_RECORD_AUDIO_HOTWORD. Ops noted by the test package are downgraded to
OP_RECORD_AUDIO. So we fix the test to account for that.

A TRANSFORMED_FROM map is used instead of a TRANSFORMED_TO one, to make
the lookup for the expected value simpler.

Bug: 186164881
Test: atest android.cts.statsdatom.appops.AppOpsTests#testAppOps
Change-Id: I5a2e621ad5f6981723465cc6469b10d72b0d9477