TextClassifier CTS improvement: wait deviceconfig propagate
We use deviceconfig to switch to CTS TextClassifierService for
testing but we don't wait for the value change before test. It should
be good to wait until the value changes before testing. We use a
DeviceConfigStateManager which waits for the value to change.

We have several flaky tests related to deviceconfig, this is the 1st
improvement to monitor the issue fail root cause.

Bug: 190889724
Bug: 190889542
Bug: 190878906
Bug: 190879071
Test: atest CtsTextClassifierTestCases --iterations 10

Change-Id: I81c81509c2f7b902c89cf8017f3d3afe5a9fffed