Use task center when tapping app in TopResumedTests
In ActivityLifecycleTopResumedStateTests
testTopPositionSwitchAcrossDisplaysOnTapSlowDifferentProcess
and
testTopPositionSwitchAcrossDisplaysOnTap tap on display content
center to bring the activity to top, but the activity might
not fill the entire display.

Solution:
Use task center points to make sure the activity
is tapped and brought to front.

Bug: 193070741
Test: run cts -m CtsWindowManagerDeviceTestCases -t
android.server.wm.lifecycle.ActivityLifecycleTopResumedStateTests
Verified on: Pixel 4 & development device

Change-Id: Ia0b8af9eb02354848ba602be6ef5ef01114b857e