Do not disable changeID for permission check as it is not supported
Shouldn't disable the change ID from the CTS, because this test apk
has targeting SDK 31(S) and the change ID undere testing should be
enabled on S going forward. This type of disable override is not
allowed by app-compat. Thus skip verifying the part when try to
disable changeID for permission check. Only verify the changeID
enable part should be good enough. If we really want to verify the
disable part, TODO: write a separate test apk with targetSDK level set to R.

Bug: 184549523
Test: atest cts
Change-Id: I39f9653793c67a66fe78d5367153442af4530b92