Support for multiple TaskDisplayAreas in ToastTest
The Test: testCustomToastBlocked_whenBehindTranslucent
does not take into account that in a multi-window environment,
new Tasks can be opened in different TaskDisplayAreas.
(With CtsActivity on one TaskDisplayArea and TranslucentActivity
on another, the Toast launched for CtsActivity is visible)

Solution:
Ensure that TranslucentActivity is opened on top of CtsActivity,
in the same Task and same TaskDisplayArea by removing
FLAG_ACTIVITY_NEW_TASK and using the Context from CtsActivity.

Bug: 192589485
Test: run cts -m CtsWidgetTestCases -t android.widget.cts.
ToastTest#testCustomToastBlocked_whenBehindTranslucentActivity

Signed-off-by: Dragos <drtigane@microsoft.com>
Change-Id: I6d1778013ad6e27ea58d4be1425ba02fc582a378