Increase timeout in LocationAccessCheckTest.
Currently a number of test runs are very flaky when the teardown method
resetPrivacyConfig() calls runLocationCheck(), which waits for a timeout
of 1 second for a job to be run. It turns out 1 second might be too
short occasionally and increasing it to 5 seconds helps it pass. There
are other sources of flakiness uncovered after this fix, but that will
be another story.

Bug: 182960931
Test: atest LocationAccessCheckTest
Change-Id: I023d487a54bfdc1f1199f14111b1eb98b1d831d4