Add sleep for DisplayHashManagerTest
There's currently know way to know when the buffer has been drawn in
SurfaceFlinger. This makes it difficult to know when to start generating
the display hash since everything must be on screen. Use sleep for now
to make sure it's been drawn. Once b/195136026 is completed, port this
code to listen for the transaction complete so we can be sure the buffer
has been latched.

Test: atest DisplayHashManagerTest
Bug: 195136026
Fixes: 192284443
Change-Id: I36cd566d3cbaa1ac71d6cc10ca7e5282fc044ca1