Reduce wait time for profiling trace file
This can reduce flakiness because if the timeout of
 sendCommandAndWaitReply(COMMAND_WAIT_FOR_PROFILE_OUTPUT)
is less than or equal to the timeout of closing file, the caller
will get TimeoutException.

It is fine to reduce the timeout for trace file because 3 seconds
are enough to write one line to file for verifying.

Bug: 192911553
Test: atest AmProfileTests
Change-Id: Id3cff55bc95f8bf0e8ed968e8212ba5c10799649
(cherry picked from commit 6483197fa1f6680d1197932fb07fef55690b149e)