ITS: Change the threshold level for R,B,G colors.
Expand the threshold value for colors to handle
any YUV processing pipeline.
For colors Black and White changed the TOLvalue to 6.
For colors Red,Blue and Green we now expect the primary_channel
value to be atleast 200 and secondary_channel max value to be
less than 55.
The VARIANCE_TOL value has been set to 17 after looking at the
existing numbers for devices.

bug: 173154486

Test: Ran test locally.
Change-Id: I07ea8a3c6c1238f73c22850fae1de8829131b158