Launch WindowUntrustedTouchTest in immersive mode
When run the test on cuttlefish_pc, text toast is located behind the
navigation, so the text toast is not clickable. As a result,
testWhenSelfTextToastWindow_allowsTouch and
testWhenSelfTextToastWindow_allowsTouch fail in the situation. Since the
test tests input touches but not window insets, launch the test in
immersive mode reduces other componnet's interference to the test goal
and resolves the issue.

Bug: 191075641
Test: run `atest WindowUntrustedTouchTest`
Change-Id: Ia559fd0980e8c8a6f8cfb069d69a41d21c047686