Fix potential flaky problem related to the SaveUi.
The test will change the EditText then click the submit button to
trigger saveui. It is possible the text change doesn't be called but
the test already calls commit(). This may cause the content doesn't
change, so saveui will not be shown.

To improve this flaky problem, the test will wait until the test
changes then call commit().

This change focus on normal text change watcher, the AntiTrimmerTextWatcher
will be work on next change.

Bug: 195371187
Test: atest PreSimpleSaveActivityTest
Test: atest SimpleSaveActivityTest
Test: atest OnClickActionTest
Test: atest OptionalSaveActivityTest
Test: atest InlineSimpleSaveActivityTest

Change-Id: Ie3771d136cbf1fa16c5b3465b84233928f8c4e0f