Fix race condition in BleClientTestBaseActivity
(cherry-pick of aosp/1793390)

During pairing in the test the onPause function unregisters the
broadcast receiver. If the Bluetooth connection is being established
before onResume is being called again, the broadcast will silently be
dropped and the test will not succeed.

Therefore change the registration to happen in onStart and
unregistration in onStop so the broadcast receiver will stay active even
when the pairing popup is still in front.

Bug: 196076707
Test: CTS Verifier: Networking->Bluetooth Test->Bluetooth LE Secure Client Test
Change-Id: I13a883a130d3f54758a4425c5aff0a48fe7707c9
Merged-In: I13a883a130d3f54758a4425c5aff0a48fe7707c9