Update input location calculation to adapt freeform window
WindowInputTests#testInjectFromThread simulates input actions to the
center of the test window. It uses half of the window width as the x
postion and half of the window height as the y position. However, if
activity is launched in freeform mode on cuttlefish pc, the calculated
position can be at the outside of the window, so the test fails
because of 'java.lang.SecurityException: Injecting to another
application requires INJECT_EVENTS permission'. Add window offset to the
input position, so that the input is within the test window even in
freeform mode.

Bug: 196924319
Test: atest WindowInputTests
Change-Id: I3516e4613d68cc635fb652bab4bbd82cbd9d1a49