Fix android.appsecurity.cts.StorageHostTest#testFullDisk fail
In the case of small resolution, the view containing the word
"internal storage" is not on the first screen, so you need to
slide the screen to find it. In APP Info, the "internal storage"
text is displayed as "xxx used in internal storage". Cannot use
scrollTextIntoView to find the full match.

Need to use scrollIntoView(new UiSelector().textContains("internal
storage")) to find the view.

Bug:197912357
Test:run cts CtsAppSecurityHostTestCases pass

Change-Id: I58a2b348868b3884060168b87fdad90cf68b0c5a