In MediaCodecTunneledPlayer, properly detect playback start
In order to detect playback start, the player would initialise the
current timestamp at 0, and check for it to become greater than
zero. This is problematic because media are allowed to use arbitrary
timestamps, and in particular some of our test media's first frames
have a 0 timestamp. For tests that check exactly one timestamp, they
would spuriously report a playback stall.

This fix uses null as a sentinel value instead of zero.

Bug: 199399807
Bug: 199394267
Bug: 187680987
Bug: 200280128
Test: atest android.media.cts.DecoderTest
Change-Id: Ie47c9a60087c2814119b8e4b456f98c244fa1faf