EncoderTest: Update timeout used in testEncoderWithFormatsParallel
Instead of using 30 seconds per run of testEncoder as a timeout, it is
now calculated based on total duration of data fed to encoders as total
duration is dependent on number of channels and sampling rate and
encoders can't be assumed to be running faster than required.

Also number of threads used in the tests is not used to divide the total
duration while computing timeout as time taken to encode in each run
can't be assumed to reduce in proportion to number of threads used
(for e.g. When running 2 instances concurrently, h/w encoders may be
twice as slower compared to 1 instance case)

Bug: 200649286
Test: atest CtsMediaTestCases:EncoderTest

Change-Id: I3e4a70c985e2523c48c9112944320337cb09f111
(cherry picked from commit 5f1b8e3d114da73de371130744e563e88a6754a6)