Fix CtsDevicePolicyTestCases for Auto.
Rather minimal changes after the previous CL fixed the infra. This
disables some known failing tests and adds a few annotations to others.

After this all tests should be passing or skipping on auto. We need to
make the setDeviceOwner change to get more running rather than skipping.

There's still some general flakiness and slowness which is being worked
on - but now auto is on the same footing.

Test: atest CtsDevicePolicyTestCases
Change-Id: Ib0dedf31b2a6b8094b0719e4733c1cbfb851900f
Bug: 200112753
Bug: 193814739