Set source crop for split screen and freeform tasks - DO NOT MERGE
This is necessary for correct rendering in SurfaceFlinger. It also lets
SurfaceFlinger know the boundary of these tasks as they may not take the
entire display.

Besides this helps devices that are freeform first handle the
reparenting when dismissing split screen because there will be a small
period of time after reparenting and task organizer setting cropping for
the new freeform state, which can lead to unexpected behaviors in
SurfaceFlinger.

Bug: 198682420
Test: testTopPositionSwitchOnTapTimeoutDifferentProcess doesn't crash
SF.

Change-Id: Ie79ff2ff5b09d9554d42eb9283691cb103467077
(cherry picked from commit 4501d44663c94329a2a3dfee49233381974973ab)