Mock modem service in CTS
A mock modem service is created to simulate the physical modem behavior.
In the CTS test, MockModemServiceConnector is used to launch and control
the mock modem service.  A test case example MockModemServiceTest is
provided to show how to connect to the mock modem service.

Bug: 203241555
Test: Build pass

Change-Id: If1f81d094561deb0eb04f66f8acf6967ed9c872c
Merged-In: If1f81d094561deb0eb04f66f8acf6967ed9c872c