Revert "Revert "mediav2 CTS: Enable codec's stop call for decoder test""
Use mediacodec stop() instead of reset(). This change was causing
problems for aac, amr decoders. Changes were made in aac, amr decoder
plugins to address this. These changes were not ported correctly causing
the decoder tests to fail. Hence this was reverted last time. This
change is now not reproducing the older seen problems. So replace reset
with stop for decoders.

Bug: 147348711
Bug: 149981033
Bug: 193194451
Test: atest android.mediav2.cts

Change-Id: I852d93f20e22b2c673b5a37e1064dde906f5ec60