Remove ListeningPortsTest from known-failures list
The ListeningPortsTest was originally added to the known-failures list
during development of the /proc/net access restrictions as the test
app no longer had access to the files necessary to check for listening
ports. However the test has since been made a host-side test, and the
test referenced in the known-failures list no longer exists under the
CtsSecurityTestCases module.

Bug: 203177211
Test: atest ListeningPortsTest
Change-Id: Id6dc7a00c9f97b2f2ad0f38cbd28487c6eb0373c