Launch test activity on virtual dipslay in fullscreen
Two tests of MultiDisplayPolicyTests are failing on cf_pc when testing
the windowing mode of activities after the activities being switched
back from removed secondary displays to the default displays. Each of
the tests launch a resizable activity in a split screen virtual display
in fullscreen, remove the virtual display, then expect the resizable
activity to be in split_window mode. However, on cf_pc, resizable window
is launched in freeform by the default, so that launching the resizable
activity in fullscreen needs to be clarified. Add
WindowConfiguration.WINDOWING_MODE_FULLSCREEN parameter to
launchActivityOnDisplay() sovles the issue.

Bug: 204233059
Test: atest MultiDisplayPolicyTests
Change-Id: I6536c497eecb192a8f95cd9d7766c6e8cb0404a4