Exclude optional algorithm tests from GSI
S launch devices can enable the ChaCha20Poly1305 algorithm for IpSec and
the framework will utilize it if the appropriate overlay is enabled.
However, when testing with GKI+GSI under R GRF, the reference GKI 1.0
image does not have ChaCha20Poly1305 (which is right: this algorithm
wasn't an R requirement), so we can't run these tests with that kernel
image, even if the vendor overlay enables the optional algorithm.

Skip the test on GSI; it will be run as part of the product image
instead.

Bug: 192113622
Bug: 203031609
Bug: 204402327
Bug: 204615046
Bug: 204860049
Bug: 233975375
Test: TH
Change-Id: I7f02aba07d37729c93b8e84fcc3fcbe6e657a0df