Speedup strict java packages tests
The bottlenecks in the strict java packages are the fetching of the jars
and extracting the actual classes, but this only needs to be done once
for all tests. Caching them means that the total time for the test case
will go down, and the per-method timings should remain relatively
unchanged.

Test: atest android.compat.sjp.cts.StrictJavaPackagesTest
Bug: 202450856
Change-Id: I63ed077e5d355379051c20d39c5658caa9354d65