Iterate on hue precision in SystemPaletteTest
- measure hue delta against median hue of palette instead of previous
hue in palette - ex. if the hue was supposed to be 250, and
when converted to sRGB for display, it was quantized to 257, and then
the next color was 243, the hue tolerance needs to be 14, not 7.

- avoid measuring hue of tones at extremes of color space
As the space terminates, at white at tone 100 and black at tone 0,
quantization  becomes more severe - ex. there are only ~12 unique
colors/hues at tone 99, and only one at tone 0.

- document quantization when converting to sRGB for display + tolerances

Bug: 205929854
Test: `atest SystemPaletteTest` run across all the wallpapers in the
solid color section to get hue coverage of new hue distance constant.
Change-Id: I69541f8da4643d71913d1a6545cd385905e378d7

Change-Id: I8458cb038fd0c34627b2db4d54f485cdb5ae2c61