Suppress RECEIVER_NOT_EXPORTED test until skip is restored in BQ
Sticky broadcasts are sent to newly registered receivers with a uid
of -1 regardless of the actual uid of the sending process. The check
in BroadcastQueue has temporarily been relaxed to allow broadcasts to
unexported receivers while this and other issues are resolved to allow
full enforcement of this feature. This commit is temporarily
suppressing the test that verifies an external app cannot send a
broadcast to an unexported receiver until these issues are resolved.

Bug: 207318980
Test: atest ContextTest
Change-Id: Iefa34283f402052e71bc980e0039c61513fe3745