Update OWNERS for ContextTest.java
AOSP is the source of truth for OWNERS files, but when attempting to
submit aosp/1899571, a merge conflict was reported when attempting to
merge this change to internal main due to the last line diverging
between the two files. Even after updating the AOSP version to match
internal main, a merge conflict is still being reported. This commit
is being created to prevent the merge conflict when submitting the
aosp commit.

Bug: 161145287
Bug: 210144408
Test: Build
Change-Id: I2315caea983a492c25f6cb78be836fbce294ff77
Merged-In: I2315caea983a492c25f6cb78be836fbce294ff77