Adjust display cutout tests with a letterbox option.
When the apps use DEFAULT or SHORT_EDGES, OEMs can have an option not to letterbox,
if the cutout in long edges overlaps at most 16 dp with app windows/contents.

If OEMs have an option to letterbox, The no cutout info is dispatched to the app.
If OEMs have an option not to letterbox, The cutout info is dispatched to the app.
Hence, some tests should be adjusted If OEMs can have an option not to letterbox.

Cherry-picked from I051ab115352ef041d8a0fda5aaa0301131fc32f7 and fixed
conflicts.

Bug: 210538874
Test: run cts -m CtsWindowManagerDeviceTestCases -t android.server.wm.DisplayCutoutTests

Signed-off-by: Sungmin <sungmin.h@samsung.com>
Change-Id: I051ab115352ef041d8a0fda5aaa0301131fc32f7