AImageDecoderTest: skip alphabetAnim.gif for some tests
In preparation for the switch to Wuffs to decode GIFs.

These tests scale down, and alphabetAnim is already small. While the
resulting images look more or less the same, the MSSIM is small enough
to be meaningless as a check. This image is still tested in other tests
in the same file, in some cases scaled down.

In addition, add the ability to save the images, hidden behind a flag.

Bug: 195936927
Test: manual
Change-Id: I6e9f78d80f7869a89bac3eebf5a68953a7158699