Make AbsListViewTest.testRequestChildRectangleOnScreen more robust
This changes this test to begin scrolled down slightly, then scrolls
back up to the top.

This avoids this test from failing in the corner case that the bottom
visible child lines up perfectly flush with the ListView bottom, to the
exact pixel.  (The reason this was failing is that
ListView.requestChildRectangleOnScreen is written in terms of the
currently visible children only, so the distanceToBottom is zero.)

Fixes: 202851341
Test: AbsListViewTest.testRequestChildRectangleOnScreen
Change-Id: I6b078e1cadb94bee8650bcc882f756c73641fbbf