Make test assertions on CDM callbacks more robust
This CL replaces waitForInvocation in favor of assertInvokedByActions
for the test callback classes. This allows to assert callback
invocations regardless of their state prior to the assertion.
There is no longer a need to explicitly clean up the internal state
of a callback before asserting invocations.

Additionally, it is now possible to expect a specific number of
invocations instead of just one.

Test: atest CtsCompanionDeviceManagerCoreTestCases
Test: atest CtsCompanionDeviceManagerUiAutomationTestCases
Change-Id: I9632ede16bb5d9f4413691f6d9ba5a5a839d0d7b