Update TextViewIntegrationTest to accommodate new toolbar architecture.
The toolbar may now be rendered by a different app than the host app.
See ag/16658303 and related CLs.
The test now uses UiDevice, and not Espresso, to find the toolbar.
Note that finding the toolbar item is sufficient to replicate this test.
This test passes for both local or system toolbars.
Other core tests relating to the toolbar will also be modified in
upcoming CLs.

Bug: 190031097
Test: atest tests/tests/textclassifier/src/android/view/textclassifier/cts/TextViewIntegrationTest.java
Change-Id: I91befb0dbf8936e0a12f0fd2f01ddebb81137066