Disable ACTION_CHARGING broadcast delay.
If the device is below 90% battery and is plugged in right before
running JS tests, the ACTION_CHARGING broadcast would be delayed by 15
minutes, and all the commands the tests ran to trigger the broadcast
would fail, resulting in a failing test. Disabling the ACTION_CHARGING
broadcast delay so that the tests can run as expected. The part of the
test that tries to trigger charging ensures the battery level is either
above 90% or there is a clear increase in battery level.

Bug: 214599394
Test: atest --rerun-until-failure 25 CtsJobSchedulerTestCases:BatteryConstraintTest
Test: atest --rerun-until-failure 25 CtsJobSchedulerTestCases:JobThrottlingTest
Change-Id: Ibae10bf7d8602bb22fb36e9a74b7eeb8fcebf869