Relocate NNAPI statsdatom tests to its own directory
Prior to this change, NeuralNetworksStatsTests existed under
cts/hostsidetests/statsdatom/src/android/cts/statsdatom/neuralnetworks.
However, because it belonged to the CtsStatsdAtomHostTestCases module,
it was not able to be marked as a "mts" or "mts-neuralnetworks"
test_suite without affecting other teams' tests.

This CL relocates NeuralNetworksStatsTests to
cts/hostsidetests/neuralnetworks under the new module
CtsNnapiHostTestCases, and marks the test as "mts" and
"mts-neuralnetworks" test_suite.

Bug: 195113404
Test: mma
Test: atest NeuralNetworksStatsTests
Change-Id: I54605080012d1835119b35e76c555011094fdf1c