Remove <Set Menu Language> CTS
This test is no longer relevant: the user needs to interact with the pop-up UI, before the language will change. This cannot be simulated in CTS and needs to be verified manually.

The tests that check that a <Set Menu Language> with an invalid language or an unexpected source are ignored, can remain. Even with the UI, these should never result in a language change.

Test: atest
Fixes: 218500277

Change-Id: Ia85382d87b710cef4317bf5c64a804c66fbd6c08