remove libbpf_android from CtsAppOpsTestCases
There's absolutely no reason to use this library from jni.

No Java process in the system has sufficient selinux privs to
actually use it.  Indeed this library is really just the core
implementation of the native boot time bpfloader process, and
it is stuffed into a library just to enable a few native tests.

The 'user side' bpf implementation (as opposed to the stuff needed
by the bpfloader) is entirely implemented inside of .h files
('bpf_headers' target), and thus there is no native library for it.

Test: TreeHugger, m CtsAppOpsTestCases
Bug: 218882781
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I4dc715918c41ebc15f6fb29075c4b75208971b4a