Perform method modifier checks in areMethodsModifierCompatible
Previously, the handling of modifiers was inconsistent. The varargs,
bridge and synthetic modifiers were handled by updating the API
description modifiers in checkMethod(). Allowing abstract to be added
to a class that has no constructors was done after
areMethodsModifierCompatible() returned false. Other modifiers were
checked in areMethodsModifierCompatible().

This change moves all the handling of the method modifiers into
areMethodsModifierCompatible() and works in a similar manner to how
they are handled in checkClassModifiersCompliance() by clearing any
modifier flags that are not significant.

Bug: 218959109
Test: atest --host signature-host-tests
      atest CtsSystemApiSignatureTestCases \
            CtsCurrentApiSignatureTestCases \
            CtsSharedLibsApiSignatureTestCases
Change-Id: I6f1dbdd2b6bc1fc61aa9f665063e4ab356846d65