Disallow mismatch caused by missing abstract keyword
Previously, it was allowed for the API to specify that a method had the
abstract modifier but the runtime did not. That was allowed for a
combination of reasons:
1. The tests compared the runtime against both the current as well as
   previously released APIs.
2. When comparing against a previously released API the checks need to
   allow for backwards compatible changes.
3. Removal of an abstract modifier that was present in a previously
   released API is backwards compatible.
4. The tests did not differentiate between comparing the runtime with
   the current API and against a previously released API.

Unfortunately, that did mean that the check against the current API was
less strict than necessary. However, now that the tests do
differentiate between current and previous APIs it is possible to
make the check against the current API more strict.

So, this change removes the ignoring of differences in the abstract
modifier on a method when comparing against the current API.

Bug: 218959109
Test: atest --host signature-host-tests
      atest CtsSystemApiSignatureTestCases \
            CtsCurrentApiSignatureTestCases \
            CtsSharedLibsApiSignatureTestCases
Change-Id: I26f1ee5e0722dcb1801eaef289efb6188ebd46c5