uwb(cts): Ensure that ranging session is closed before exiting test
The ranging session test should ensure that the ongoing session is fully
closed before exiting to ensure that other tests in the suite are not
flaky. Use a separate latch for the ranging reports callback to avoid
races between report callbacks & stop/close callbacks (both of them can
occur around the same time when the session is ongoing).

Bug: 196225233
Test: atest UwbManagerTest
Change-Id: I6977a27d7b3cd1b38996f1c19343cbef10df740f