Refrain from persisting shortcuts into AppSearch in low end device
As observed in b/220133968, on low-end devices, persisting shortcut into
AppSearch on a background thread introduces 50ms delay on app start time.

Given that there isn't a clear product roadmap for such feature to work
on these devices, we will disable the integration until we have a concrete
demand for such feature that justfies the performance cost on low end
devices.

Bug: 220133968
Test: atest CtsShortcutManagerTestCases
Change-Id: I40f7d41609551f62843bee8ca28ce1adb6db54de