Fix screen brightness
Remove 140 as a tested value since it is too high.
The proper way to fix this is probably to test values that are below the
current brightness. However, this would require passing the values to
test to the device, which is a bigger refactor. This is a quick fix that
should make the test pass.

Test: atest UidAtomTests#testScreenBrightness
Bug: 221803371
Change-Id: Id0f44741bab2a81efa3bccb1840fabe583e73696