Improve stability of GameService CTS tests
Fix flakiness in
takeScreenshot_expectedScreenshotSaved by
waiting until screenshots are available
on disk before running assertions. Also
show the overlay by default to avoid any
need to interact with the system bars
in order to show the overlay. This
makes the tests simpler and removes the need
to synchronize as many different steps.

Bug: 223981361
Test: atest CtsGameServiceTestCases
Change-Id: I3fb04cd4b7e24f463ce6dfef7672538db44491eb