Delay opening the camera and mic more when activity starts
This will make the test more robust to device performance regressions
not related to the test. These tests would sometimes fail because they
would open the sensor before the activity was fully drawn on the screen
which broke the assumption in the test that the activity would be
visible first.

Also try moving all of this into onStart

Test: atest CtsSensorPrivacyTestCases:android.sensorprivacy.cts.SensorPrivacyCameraTest#testTapjacking --iterations 100 -- --abi x86_64
Fixes: 224887714
Change-Id: Ie5a0a174eaaef331eb3186893c6e8d7e611a8cd5