Fix QuickAccessWalletClientTest.
For the case where you send an event, onCreate is called but not onBind.
ag/16980283 removed the override of onCreate where we store a reference
to the service instance, so we never sent any events.

That broke QuickAccessWalletClientTest.
Putting it back fixes the test without breaking anything else, though
the test is still a bit flaky.

Fixes: 225066814
Test: atest QuickAccessWalletClientTest --retry-any-failures
Test: atest QuickAccessWalletServiceTest
Change-Id: I10b30ba49c3f5e26f3bd69994c36132ec3715397