Correct KeyAgreementTest assumption about ECDH security level
A test case was added to ensure that ECDH keys are generated in hardware
for KeyMint. This is only a valid assumption if the underlying KeyMint
implementation is itself hardware.

Fix the test by checking that ECDH keys have the same security level as
ECDSA keys. If one is in HW, then both should be in HW. If the underlying
KeyMint is SW, however, then we should allow them both to be SW.

Test: KeyAgreementTest
Fixed: 223401378
Change-Id: I44e6944ae6b9e23268959ae91edd1bd140d3349c