DO NOT MERGE Add new privileged permission for unique id attestation
This permission will eventually replace the sepolicy that we have now,
making it possible to test unique id attestation, which is no longer
possible due to the deprecation of shared uids.

Skipping automerger because the framework manifest has diverged too
much across aosp and tm-dev to allow for clean auto merges, and
presubmits block.

Test: android.permission2.cts.PermissionPolicyTest
Bug: 216778747
Change-Id: If59fd7f20d2424ef6275e32ce29631d27ff7adc7