Add tests for new API to disable sending touches to wallpaper
Currently, all touch events on a Window are sent to wallpaper,
if the window uses FLAG_SHOW_WALLPAPER to show system wallpaper
in the background. This may have a security implications if
the Window is showing sensitive information. So, we are adding a
new API to allow developers to block touched to wallpaper.
This CL adds tests for testing the new behavior and API.

Test: atest android.server.wm.WallpaperWindowInputTests

Bug: 177054831
Change-Id: I881226dea94bee68b92c1f15a71e0e8f528160da