signature tests: Memoize access to info derived from instrumentation args
This change memoizes access to information derived from instrumentation
arguments as static Supplier objects.

Instrumentation args are provided to the whole process not just an
individual class so it is safe to memoize information derived from
those arguments at the class level. Memoizing all access means that
the test does not need to enforce a specific initialization order.
This will be important in follow up changes which will use the memoized
information outside the normal flow of execution within a test.

Bug: 228322252
Test: cts/tests/signature/runSignatureTests.sh
      atest --host signature-host-tests
Change-Id: Ia5c55b87d0fe219fbf6e0e0f583d557ec4ed49ab