Add new test for writing to an AHB from a compute shader
The ability to use AHardwareBuffer-backed images as read-write resources
in a shader is a new optional feature in Android T. This test provides
some initial light coverage for it.

If the device supports using an RGBA8 AHB as a storage image, this test
uses a compute shader to write a test pattern into the image, and then
maps the AHB on the CPU to verify the contents.

More extensive testing should be done in dEQP instead.

Bug: b/203537680
Change-Id: Iaa54968b60444d9181778d234ea7694930753335