CtsCameraTestCases: Fix failure in testCamera2AccessCallbackInSplitMode.
This failure was caused by the remote activity process ending due to a
NullPointerException. The AvailabilityCallback and StateCallback objects
stay around for longer than the activity (after onDestroy is called).

By handling screenLayout config changes, the activity won't be destroyed
abruptly. In addition to this, I have moved the callback objects onto a
handler thread (this is the recommended way of doing things).

Change-Id: Ic168e510bffaf036f4568e77bddf544e06e2f4a2
Test: Ran testCamera2AccessCallbackInSplitMode.
Bug: 228778561