Fix RecognitionServiceMicIndicatorTest
Updates the car privacy dialog content id so that the dialog
and the list of app labels can be found by the test. Also, removed
separate assertion for cars that checks that the app is blamed
regardless of the trustedVoiceService parameter, but there doesn't
seem to be a need for this logic.

Bug 225292921

Test: atest RecognitionServiceMicIndicatorTest

Change-Id: I9b4f57a6a6f9308374de6b62afb5b181289798d1