add button to advertise device
The app now has a button to start advertising the device so that another device can associate/pair with it. Previously, this was done by going to the Settings app (on the "Pair New Device" screen).

The app will advertise the device for 100sec, which is arbitrary.An optimization we could consider is to stop advertising it after the device has been paired.

In order to advertise a device, an app needs to hold the BLUETOOTH_CONNECT permission. This is not indicated in the documentation. Since it's a dangerous permission, it needs to be explicitly granted by user (see screenshots for reference)

This CL also removes some unused permissions that were remnants of the legacy test app.

Screenshots:
- https://screenshot.googleplex.com/BLQTdKXZg3sZwX8
- https://screenshot.googleplex.com/9aMUuRhS6cw4aSV
- https://screenshot.googleplex.com/Ao86QBePAJikh65


BUG: 230382800
Change-Id: Icf15eac5d1b39f7146b1fe6464b3b1a806ff4d66