Fix testShowInputMethodPicker failure
UiDevice.getInstance(Instrumentation) return a singleton
value if it's already existing. The cached value somehow
fail to detect the MOCK_IME_LABEL.

Add a workaround to clear the internal singleton value
which make UiDevice create a new instance and a TODO bug
to track we should remove this workaround in the future.

Bug: 230095307
Bug: 230698095
Test: atest CtsInputMethodTestCases
Change-Id: I160309647fd0e63ae270ed6298e1e7523aa0aaf6