Reland: Update ANGLE CTS tests for when ANGLE is system driver
Some tests failed with the previous version of this change on devices
that use ANGLE as the system driver, but have no legacy driver.  Once
the "ro.hardware.egl_legacy" system property lands, it becomes
possible to distinguish devices that have only ANGLE, and so the
appropriate tests can be run or skipped.

Test: CtsAngleDeveloperOptionHostTest
Bug: 224558229
Bug: 230616681
Change-Id: I03483f1c5ec6ed17cc9c34cabf11c45ef07debc5