Ask MediaProvider if FUSE passthrough is enabled
Relying on the persist.sys.fuse.passthrough.enable system property to
check if FUSE passthrough is enabled is not a reliable solution, as
shown by mixed builds with GSI. The problem arises when an underlying
system not supporting FUSE passthrough (e.b., Android R) runs with an
Android S+ GSI which wants to use FUSE passthrough, thus has
persist.sys.fuse.passthrough.enable set to true.
Android works fine as MediaProvider (FuseDaemon) catches the exception
that the kernel does not support FUSE passthrough, but the Android S+
tests assume that FUSE passthrough is enabled because of the system
property.

With this change, the tests ask directly to MediaProvider if the FUSE
passthrough feature is being used or not, to avoid this kind of
inconsistencies between MediaProvider and tests with mixed builds.

Bug: 207733510
Test: android.scopedstorage.cts.device.ScopedStorageDeviceTest
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: Idfdb1d99c3da5cdc20c56e0234543e382cf72689
Merged-In: Idfdb1d99c3da5cdc20c56e0234543e382cf72689