EncodeVirtualDisplayWithCompositionTest: handle encoding errors gracefully
Previously upon encoding errors, the encoding thread would call junit
fail() which crashes the thread and in turn the whole test process,
preventing subsequent tests from running. Only the main thread is
supposed to call fail().

To fix this, we save the error from encoding thread and raise it on the
main thread in stopEncoding().

Bug: 228828942
Test: android.media.cts.EncodeVirtualDisplayWithCompositionTest#testVirtualDisplayRecycles
Change-Id: I1bd42c6c86a1c9f91d83e95a57f22cfa610e5002