Dump config in onMultiWindowModeChanged for large displays
On a device with a very large display, even when an activity
transitions from split to fullscreen, no configuration change
happens as it may not cross any of the layout boundaries.

As onMultiWindowModeChanged is guaranteed to be called when an
activity exits split, let's just use this to dump configuration.

Bug: 233146813
Test: atest android.server.wm.AppConfigurationTests
Change-Id: I883f3c05db65fd8d17175bda37138bbe0b2cc11a