AppDataIsolationTests: remove useless code for emulated FBE
testDirectBootModeWorks() contains code that tries to enable emulated
FBE.  However, it only runs if emulated FBE is already enabled, which
makes it useless.  Moreover, the whole test method is skipped on devices
that aren't using native FBE, which makes it doubly useless.

And in any case, emulated FBE is obsolete and is planned to be removed.

For all these reasons, remove this unneeded code.

Test: atest AppDataIsolationTests # on Cuttlefish
Change-Id: Icdd9a598dd38582b3705d7b74e126117c88505a1