ShadowCallLogTest: remove useless code for emulated FBE
testDirectBootCallLog() contains code that tries to enable emulated FBE.
However, it only runs if emulated FBE is already enabled, which makes it
useless.  Moreover, the whole test method is skipped on devices that
aren't using native FBE, which makes it doubly useless.

(Technically, the test method ran on devices that used either native or
emulated FBE.  But in practice, emulated FBE would never already be
enabled at the beginning of a CTS test.)

And in any case, emulated FBE is obsolete and is planned to be removed.

For all these reasons, remove this unneeded code.

Note, the whole test class is actually marked with @Ignore, so it's not
running anyway.  But I'll assume that will be fixed at some point.

Test: atest ShadowCallLogTest # on Cuttlefish, after removing @Ignore
Bug: 232458753
Change-Id: Ib88e9a3f226b25bf6bb2731a3cc0e4a7090e4ac7