Add dynamic-config to CtsCurrentApiSignatureTestCases
This is required to enable the CTS 12 current api tests to pass on
Android 12 devices with the Android 13 ART Module installed. A small
number of changes were made in the Android 13 timeframe that changed a
few APIs in (hopefully) inconsequential ways. API's are now frozen so
these can not now be corrected.

Earlier work enabled the use of dynamic-config in other signature
tests (b/177303736).

Bug: 177303736
Bug: 233719013
Test: atest CtsCurrentApiSignatureTestCases
Change-Id: Ie0549990e1e250e86b0cd8a9318ba671b2f4ddbe