mediapc: getRequiredMinConcurrentInstances cleanup
checkAndGetMaxSupportedInstancesForCodecCombinations() used to call
getRequiredMinConcurrentInstances() to query minimum number of instances
that are required to be supported.
This function was coded as per 720p requirements alone and doesn't handle
1080p requirements correctly. So instead of calling this function,
checkAndGetMaxSupportedInstancesForCodecCombinations() now takes minimum
required instances as an argument.

Also requiredMinInstances passed to testCodec() in transcoder tests
was wrongly halved.

As per media performance class requirements, 6 instances of
transcoder sessions (i.e. both decoder and encoder running
concurrently) are expected to be supported. In earlier versions
this requirement (for both instances and achieved fps) was halved
as transcoder involves both encoder and decoder and such an approach
was followed in other tests that measure performance.
But in media performance requirements, this halving is wrong.
This is now fixed.

Bug: 228104072
Test: atest CtsMediaPerformanceClassTestCases:MultiDecoderPairPerfTest \
 CtsMediaPerformanceClassTestCases:MultiDecoderPerfTest \
 CtsMediaPerformanceClassTestCases:MultiEncoderPairPerfTest \
 CtsMediaPerformanceClassTestCases:MultiEncoderPerfTest \
 CtsMediaPerformanceClassTestCases:MultiTranscoderPerfTest

Change-Id: I56ed232faddf823ad91c3bae6ee79de51e65c175