Specify launch TDA in ActivityVisibilityTests
Some of the tests from ActivityVisibilityTests class verify or depend
on Home Activity lifecycle but the tests don't take into account the
the possibility of Activities being launched on different
TaskDisplayAreas.

Change the tests to use the new
ActivityOptions:setLaunchTaskDisplayAreaFeature API when launching the
test activities to make sure they are launched into the same as the Home
Activity.

Also remove the assumption for the same TaskDisplayArea as it becomes
redundant with this change.

Bug: 234349363
Test: atest ActivityVisibilityTests
Change-Id: I619ebd51aaff5e069354a7fc18b068c49c00271e