Specify launch TDA in Intent Tests
Some tests in this class start multiple tasks and assume they will end
up on the same display area. This is not always the case for devices
with multiple screens.

Change the tests to use the new
ActivityOptions:setLaunchTaskDisplayAreaFeature API when launching the
test activities to make sure they are launched into the same task
display area.

Bug: 232437966
Test: run cts -m CtsWindowManagerDeviceTestCases -t
android.server.wm.intent.IntentTests

Change-Id: I032e67c94fa832a5de35099029db7074037f5bf0