Skip new RS tests under instant mode.
Some RS tests checks the vendor API level to determine if it should be
skipped. However, instant apps are denied access to the system properties,
and the utility method used to get the vendor API level would fail. This
patch ensures such tests are only exercised in full app mode.

Bug: 232489547
Test: CtsRenderscriptTestCases
Change-Id: I9832c3d3e4641f10e9f98385b9da89b88d660978