Check FEATURE_ACTIVITIES_ON_SECONDARY_DISPLAYS in virtual input/audio tests
This flag must be set to true in order to allow activities to run on
non-default displays. Since virtual input and audio devices are only
valid on a VirtualDisplay (associated with a particular
VirtualDevice), the tests for these APIs are only valid when the flag
is set.

Bug: 233712769
Bug: 236208910
Test: atest tests/tests/hardware/src/android/hardware/input/cts/tests/Virtual*Test.java VirtualAudioTest
Change-Id: I33e85c234296216083fd177243c991db0d49166e
(cherry picked from commit 783d87c45cb8a342ea03f3e2993df277ef4bce91)