Specify launch TDA in ActivityMetricsLoggerTests
Affected test: testAppRelaunchSetsWaitResultDelayData

The test was launching 2 tasks and assumed they will be launched on
the same display area. When a "bring-to-front" intent would be sent,
the expected behavior would not occur because the first activity
would still be running and receive that intent and do nothing.

Change the test to use the new
ActivityOptions:setLaunchTaskDisplayAreaFeature API wghen launching the
second activity to make sure they are launched into the same TDA

Bug: 235525919
Test: atest CtsWindowManagerDeviceTestCases:ActivityMetricsLoggerTests#testAppRelaunchSetsWaitResultDelayDat

Change-Id: Ibd0148bc622c79fcddd516b55abb6342affbfa09