VirtualDeviceTestCases: Reintroduce sleep to reduce test flakiness
While the sleep is no longer necessary for VirtualDevice API
correctness, we're seeing a substantial increase in test flakiness due
to a longstanding issue with extraneous hover events being generated
when the sleep was removed.

We'll temporarily reintroduce the sleep here until we can debug the
underlying issue with better debugging tools.

Bug: 230009847
Test: atest VirtualMouseTest
Change-Id: I46eeee7cbaf0c66c7b26b0dbca09107bccf5462d