Create and delete test clone user once
The clone user is created once before all tests run
and is deleted after all the tests run. This aims to
minimize the duration that CtsAppCloningHostTest runs for.
The tests run significantly faster with these changes.

Here is a report with a comparison of the different run
times: go/ac-test-comparison-times

Bug: 230733397
Test: atest CtsAppCloningHostTest
Change-Id: Ic593836ed95a5301deed74558ca8fcaca3409f80