Fix AccessibilityWindowReportingTest#showNotTouchableWindow_activityWindowIsNotVisible
This test is failed if the device is under 3-botton navigation mode.
The failure is due to that the non-touchable window doens't overlap with
navigation bar window, and the touch region of the navigation bar window
is not consistent with its window frame.
To fixt it, we set the inset types to 0 to make the non-touchable
window full-screen.

Bug: 236595250
Test: Change device to 3-botton navigation mode and run
      atest AccessibilityWindowReportingTest
Change-Id: Ic4d007d15bb733f552c852e44f8f381156348ee8