Add CTS test to exercise Camera -> Graphics Shader -> Encoder path
We need a basic cts test which has an end to end flow from capturing
the video, applying graphics filters and passing it to the encoder. For
this test, we are using a fragment shader to tweak the color of the
video every other frame.

This test does not put bounds on the performace, i.e we don't check the
encoding/decoding logic, or the frame drop. We only check for the
common GL errors which may occur.

Bug: 184858824
Test: The CTS test added passes within a reasonable time limit.

Change-Id: I4ec35953c7814995a5dd42bf8ef958317095d094