Test that accessibility focus is treated as keep clear area
This test installs & enables a test accessibility service, then puts
accessibility focus on a (otherwise unfocusable) View, and tests the
activity has a keep clear area matching the View's bounds.

The accessibility services requires flagRequestTouchExplorationMode, as
accessibility focus is only set if touch exploration mode is enabled.
Every time `getUiAutomation(flags)` is called on the Instrumentation,
the flags are applied. By default AccessibilityServices are suppressed
if an UiAutomation is active, which causes issues with the test as
accessibility focus does not exist if no AccessibilityService is active.
WindowManagerState calls getUiAutomation() from executeShellCommand in
computeState. An added flag to WindowManagerState specifies whether
UiAutomation should suppress accessibility services.

Bug: 220108592
Test: atest KeepClearRectsTests
Change-Id: Icb27e1a100c7267b7d2864a0a01d4f7404dcb0bf