Camera: Assert no onTorchModeChanged() after camera open
If flashlight is on when camera is opened, the camera device shouldn't
call onTorchModeChanged().

The reason is that if onTorchModeChanged(OFF) is called, it could race
with the onTorchModeUnavailable() triggered by camera framework, thus
confusing the application. For example, application may first receive
onTorchModeUnavailable() immediately followed by
onTorchModeChanged(OFF). In this case, the application thinks that the
flashlight can be turned on, but in reallity it couldn't.

Test: Run FlashlightTest#testCameraDeviceOpenAfterTorchOn
Bug: 237396927
Change-Id: I591af8f47b399762f5055c9c1353d0a757f1c822