Ensure to set inset animation callback to null
In WindowInsetsAnimationControllerTests, TestActivity is launched
in the set-up process but this activity has window animation
callback by default. If some insets change when the activity is
launched, some inset animation starts with this default inset
animation callback.

In each test case of WindowInsetsAnimationControllerTests,
the animation callback is replaced with VerifyingCallback for the
same (already-running) animation. In this case onStart() is called for
the default callback while onProgress and onEnd() are called for
VerifyingCallback, which is not expected by VerifyingCallback.

As it's up to each vendor whether to change some insets when an app is
launched, the root problem here is that callback is replaced during a
test. This CL ensures that no callback is set at the beginning of each
test.

Bug: 206335742
Test: atest --rerun-until-failure 100 WindowInsetsAnimationControllerTests
Change-Id: If8e22ce5663ff40c2cc96666d4fcb0edafcf69cd