Get the surface for camera preview on each test iteration
Because of changes specific to A12 it can happen that the test activity
gets recreated because of size configuration changes. Because the
activity gets recreated, the initial surface will be invalid for the
next iterations.

Having this possibility in mind, it would be safer to get the surface
from the activity on each iteration - to take into account the possible
changes that might appear.

Bug: 238399415
Test: atest CtsMediaStressTestCases:MediaRecorderStressTest
Change-Id: I053657e1605a6a60a9e6e0ff95969bcd967d06f6