Simplify InputMethodPickerTest
This is a follow up CL to our recent CLs [1][2], which introduced a
regression test for Bug 181660568 and Bug 236101545, respectively.

InputMethodPickerTest already asserts that the IME picker is not
visible in its @Before method.  Hence we do not need to verify it
again in each test.  This CL just removes such redundant assertions.

Also, @AppModeFull needs to be placed at the test class is better to
be placed at the InputMethodPickerTest class level because all the
test methods defined in this class will try to send
ACTION_CLOSE_SYSTEM_DIALOGS because of the @Before method.  Hence this
CL moves @AppModeFull to the class level.

This CL also simplifies several lambda expressions with method
references, with no behavior changes.

Lastly,

  assumeTrue(mContext.getPackageManager()
          .hasSystemFeature(FEATURE_INPUT_METHODS));

is better to be put in @Before method as verifying the IME picker
visibility would make sense only when installable IMEs are fully
supported.  Form factors that do not support FEATURE_INPUT_METHODS
should be responsible for their custom text input experience by
themselves.  This CL simply moves it to setUp() method.

 [1]: Id17ff13bebe88ac8ff07efe0155fe7acc3470c97
      75acc2860b1671dec98fce9454df21e70e2a17ef
 [2]: Id0f7213ae2be5a6c75212f9407e550c046b8c506
      103ed4e4fb94be8690b347c72254b093f17d2527

Bug: 234882948
Test: atest InputMethodPickerTest#\
        testShowInputMethodPicker_noDismissWhenOverlayPopup
Change-Id: Id9c84110a86c78bbfaed132b01c83dd1339def5d