Adding pop-up toast notifications for User Consent for Device Logs Test.
This change list adds 2 pop-up toasts for both test cases upon
completion. Before this change the test does not notify user that the
test was successfully completed. User was only notified when the tests
failed.

Along with adding the notifications, there were minor clean ups
including removal of unused constants. And changing one of the executors implicit runnable
functions to lambda to make it consistent with the other test method.

Bug: 236969239
Test: Manual deployment and test on virtual device and on a physical
bramble device.

Change-Id: Ie9bd4efcf5b94bda6b601b5bb8ec440515f4bea0