Add missing test dependencies for CtsAppOpsTestCases
In Android.bp, all dependencies of a test should be declared explicitly, so that test infrastructure could only provide the needed dependencies instead of all to run a test. This change adds the missing dependencies that are used in test config file but not declared in Android.bp.

Bug: 222438068
Test: atest-src CtsAppOpsTestCases \
          --experimental-device-driven-test \
          --experimental-java-runtime-dependencies \
          --bazel-arg=--test_timeout=48000
Test: atest CtsAppOpsTestCases
Merged-In: Ica6f734a8dc9284c8ca1ac4b8b673b42cf4e7a4f
Change-Id: I34d4b6b3f7fbaa4b0cea9a2898bfe54c0b8c8b8d