Rename and re-order tunnel mode tests
Rename the tests to better align with what the test is doing, also
update the comments and re-order them so all tunnel mode tests have the
same layout.

Also, no need for SDK version suppression, since the tests were
introduced in the release that the suppression is for.

Also, reference which APIs the tests add coverage for.

Bug: 220266381
Test: build all tests
Change-Id: I672d60d7b7b97165b045774e3e00e1128c9a9c5f