Add a missing test dependency for CtsVoiceSettingsTestCases
In Android.bp, all dependencies of a test should be declared explicitly, so that test infrastructure could only provide the needed dependencies instead of all to run a test. This change adds the missing dependency that is used in test config file but not declared in Android.bp.

Bug: 222438068
Test: atest-src CtsVoiceSettingsTestCases \
          --experimental-device-driven-test \
          --experimental-java-runtime-dependencies \
          --bazel-arg=--test_timeout=48000
Test: atest CtsVoiceSettingsTestCases
Merged-In: I307369f7e3ac72b9cabab5c8aa635bcef009d737
Change-Id: I6c2ada6ac50f32ae62712a7f9b1df22ef641d1de