Add missing test dependencies for CtsWidgetTestCases
In Android.bp, all dependencies of a test should be declared explicitly, so that test infrastructure could only provide the needed dependencies instead of all to run a test. This change adds the missing dependencies that are used in test config file but not declared in Android.bp.

Bug: 222438068
Test: atest-src CtsWidgetTestCases \
          --experimental-device-driven-test \
          --experimental-java-runtime-dependencies \
          --bazel-arg=--test_timeout=48000
Test: atest CtsWidgetTestCases
Merged-In: I307369f7e3ac72b9cabab5c8aa635bcef009d737
Change-Id: I5e642fbbc5db05948a74458c00f9ac3d9c9e3b98