Add missing test dependencies for CtsScopedStorageDeviceOnlyTest
In Android.bp, all dependencies of a test should be declared explicitly, so that test infrastructure could only provide the needed dependencies instead of all to run a test. This change adds the missing dependencies that are used in test config file but not declared in Android.bp.

Bug: 222438068
Test: atest-src CtsScopedStorageDeviceOnlyTest \
          --experimental-device-driven-test \
          --experimental-java-runtime-dependencies \
          --bazel-arg=--test_timeout=48000
Test: atest CtsScopedStorageDeviceOnlyTest
Merged-In: I654a4e3b19dedd3d1e172774c5fcbc84146d02f7
Change-Id: I207cf6397fd1d0a2b90b196be01b3ed43e57f8fc