Test that sensitive information is stripped from ActivityOptions
The implementation of shared-element transitions takes the
ActivityOptions from the calling activity and sends them to
another activity. This means that any sensitive information
passed into ActivityManager via ActivityOptions can make its
way to an unrelated app.

This test makes sure that the sensitive information is stripped
out of the ActivityOptions that gets sent to the target app.

Bug: 237290578
Test: testActivityManager_stripTransitionFromActivityOptions
Change-Id: I6acf40c8da70649af0315429fdca705fa804b0bb