Add `cts-dalvik-host-test-runner` for CtsIcuTestCases
The `cts-dalvik-host-test-runner` is required because the corresponding TF test config uses the `LibcoreTest` class which is defined in `cts-dalvik-host-test-runner` to run the test.

All of a test's dependencies should be explicitly declared in the Soong build module to enable infrastructure features that ensure correctness and optimize runtime. So that this change declares this dependency in the `host_required` attribute.

Bug: 220991354
Test: atest-src CtsIcuTestCases \
          --experimental-device-driven-test \
          --experimental-java-runtime-dependencies \
          --bazel-arg=--test_timeout=48000
Test: atest CtsIcuTestCases
Change-Id: Iecce0116f6d8292a673426cdeb185ab6503f1b08