Add several Nene benchmarks and run them with JUnit via a custom BenchmarkTestRunner.
Fixed an issue with FrameworkMethodWithParameter where equality and hashcode was causing different tests to appear to be the same test from JUnit's perspective.

Benchmark results are currently logged with the log tag 'MetricCollector', however MetricCollector can easily be modified to report the results as a text file instead, which will be done in a future CL.

The system is unstable when running these benchmarks and usually crashes before all of them can run in succession, due to various errors, such as the system server crashing, and out of memory exceptions.

Change-Id: I7a611c59fe9e00be5ca75b2cfcc18ce5e1052964
Test: atest cts/common/device-side/bedstead/nene/src/test/java/com/android/bedstead/nene/benchmarking/NeneBenchmarksTest.java -- --module-arg NeneTest:device-listeners:com.android.bedstead.nene.benchmarking.MetricCollector
Ignore-AOSP-First: Working in Gerrit for this project.