MediaV2: Avoid passing Android Version U to ApiTestUtil methods
Functions in ApiTestUtil are being updated to throw an exception
when they are called with android versions that are not finalized
yet (i.e. if the versions are defined as CUR_DEVELOPMENT)

Also cleaned up calls to ApiTestUtil methods to use Android T where
appropriate.

This is a modified cherry pick of what is in master, where development
is already underway for the next big release.

Bug: 223868241
Bug: 242600628
Bug: 242335889
Test: cts/mts CtsMediaCodecTestCases/android.media.codec.cts.MediaCodecTest
Merged-In: Id9db89f686362a04df15e293966fc60c767789ea
Change-Id: I147e042c82cd327b9b8787804b7bf418cc319407