fix CtsVerifier test where call gets stuck
A partner reported a failing CtsVerifier test.  Upon investigation,
it looks like a stuck call at the start of the test was causing the
test to fail.

To remedy this, internal cleanup related was done. Additioanlly, the
test cleans up possible connections at the start and end.  A delay was
also put in the test to avoid edge case bugs.

bug: 239486526
Test: 1 new unit test,
      1 edit incoming SM CtsVerifier test
Change-Id: I8302b07a0e56bd8218d369874f750d32c89ae7d2