Remove the workaround for TextViewIntegrationTest
There is a stale "NotificationShade" window from SysUI stealing the
inputs. The window is in the "exiting" state and seems never finish
exiting, this is a workaround to reset the state. Because we only
meet the issue in mts test invocation and this test is removed from
mts now. We don't need this workaround that may break the test if
the notification doesn't collapse or the action is slow.

Bug: 244210430
Test: atest TextViewIntegrationTest
Change-Id: I446598e3c818380e235be5e2d038d7ae0070070d