Added missng INTERACT_ACROSS_USER permissions to some tests.
Both AssistUtilsHelperTest and ActivityManagerHelperTest need this
permission but were not explicitly granting; yet somehow they used
to get it because of UserManagerHelperTest, which was calling
UiAutomation().adoptShellPermissionIdentity(CREATE_USERS)
in a @Before method.

But then UserManagerHelperTest was refactored to use DeviceState
(on commit 42c39d195d0d4628d2c00e483f0fe14db7e8be96), and puff! - the
granted permission is gone!!!

Test: atest AssistUtilsHelperTest ActivityManagerHelperTest

Fixes: 246891403
Fixes: 246891213

Change-Id: Ief811928a3c4cf3d2df8616c7d2bdb8ad5fbc676