Add `cts-dalvik-host-test-runner` for CtsLibcoreWycheproofBCTestCases
The `cts-dalvik-host-test-runner` is required because the corresponding
TF test config uses the `LibcoreTest` class which is defined in
`cts-dalvik-host-test-runner` to run the test.

All of a test's dependencies should be explicitly declared in the Soong
build module to enable infrastructure features that ensure correctness
and optimize runtime. So that this change declares this dependency in
the `host_required` attribute.

Bug: 220991354
Test: atest CtsLibcoreWycheproofBCTestCases
Change-Id: I5f0c4a235f52f976f3a155a77afcaf226da7415e