Only verify the first 2 events in VirtualTouchscreenTest#sendTouchEvent.
There's an extraneous ACTION_HOVER_ENTER event in builds before T QPR1. We fixed the related bug (b/244744917) so the hover event is not generated anymore. In order to make the test permissive to existing and new behaviors we only verify the first 2 events here. In U we should continue to use verifyEvents().

Bug: 250759209

Test: atest android.hardware.input.cts.tests.VirtualTouchscreenTest#sendTouchEvent


Change-Id: Ic09a0c9ed0dc38d78236bc21ec5b7a2c7ad9f8ad