Polling for wifi network after wifi disabled.
In the wifi cts test for Wifi getCurrentNetwork API, after disabling
Wifi, the getCurrentNetwork may not immediately return null. The Wifi
state machine needs to stop the client mode manager, and clear the
current network. The current network is not guranteed to be cleared
when the Wifi get connection returns network id -1.
Needs to poll and wait for the network to be null.
Tests both disconnet network and disable wifi cases.

Bug: 252814919
Bug: 254145509
Test: atest android.net.wifi.cts.WifiManagerTest#testGetCurrentNetwork
Change-Id: I3592ed79b30964c83ccd06a6b104f53d257af0d5