Mark mPendingRequestToken volatile
This value is always changed in the same thread within a single test.
The other thread only reads its value, so as long as we ensure
immediate visibility to other threads, atomic assignments and
happens-before to ensure mPendingResponse is ready to take new
responses, it should be good enough. Therefore I didn't make it an
AtomicLong.

Note different tests are running in different threads because they have
time out set up. In general tests are still sequential unless one of
them times out. If that happens then future tests that use the same
session as the timed out test may fail, but that's out of the scope of
this fix.

Bug: 254712131
Test: Presubmit.
Change-Id: I3a95382bd24d1072058fe8fc886542d038f4c948