Fix CTS testTrimMemActivityFg on Wear
The activity SimpleActivity is used in both packages com.android.app1
and com.android.app2.
Enabling swipe-to-dismiss on Wear has visibility and impacted
life-cycle during this test.

After launching the app2, this test expects that the process state of
the app1 should be changed to PROCESS_STATE_SERVICE, which is not in this test.
This change disables swipe-to-dismiss for the activity SimpleActivity on Wear.

Bug: 250685635

Test:
 run cts --html-in-zip --skip-all-system-status-check
   --skip-device-info --skip-preconditions -m CtsAppTestCases
   -t android.app.cts.ActivityManagerTest#testTrimMemActivityFg

Change-Id: I05483368ccce7c9a9754a24ad70c62202dc34346