Update WebViewTest to JUnit4
We are going to share WebView tests between
SDK Sandbox tests and regular WebView
tests.

This means each test suite will need to inherit
from a shared class.

This means that the WebView tests
can no longer make use of
ActivityInstrumentationTestCase2.

We are updating WebViewTest to make
use of JUnit4 while we are at it.

This will make it harder to back merge test fixes
to previous CTS versions in the future
but it is worth it because it means
we will be able to
share tests easier with the SDK Sandbox.

It should also be a little easier to copy
these tests into
AndroidX since they are now running in the
same environment (previously using different
junit versions).

Finally formatted the tests with `google-java-format --aosp`
since we would already have trouble cherry picking in the
future.

Test: Confirmed that the instant app tests
are still filtered out. Purposefully made some tests
fail to ensure that they would be reported correctly.
Modified NullWebViewUtils.isWebViewAvailable to return
false to ensure that it showed tests have failed
assumptions. Also confirmed that the same number
of tests were seen to run (77).

Bug: 1377308
Change-Id: I58a8246fce2496ac4411c07b8d8393d0af1ea4bf