BroadcastQueue: more misc fixes before dogfooding.
When sending broadcasts destined for a mix of registered and manifest
receivers, we need to send the combined list together to avoid the
second duplicated send from applying BroadcastOptions policy to skip
the registered receivers sent moments earlier.

SystemActions needs to send its a11y broadcasts as "foreground" since
they're user-requested actions, otherwise they risk being delayed
beyond the short timeouts used in tests.

PackageManagerShellCommandTest needs to wait for a barrier in the
queue, as it can often get backed up by more than the 2 second
timeout that it's willing to pause for.

FocusHandlingTest needs to send its broadcasts as "foreground" since
it expects the fake IME to input data quickly.  We normally avoid
delaying broadcasts for apps under instrumentation, but the fake IME
looks like a separate unique app not connected directly to the test.

Bug: 254769026, 254763137, 254761140
Test: atest PlatformScenarioTests:QSMediaController
Test: atest CtsInputMethodTestCases:FocusHandlingTest
Test: atest CtsContentTestCases:PackageManagerShellCommandMultiUserTest
Change-Id: I58b84e46a24cfc926cf0498c5c1957c5d382369b