CtsMedia: Fix arguments to InputSurface created in clearSurface()
aosp/2168786 extended InputSurface() class to test 10-bit decoding
and while doing so had overloaded InputSurface(Surface) constructor
to pass useHighBitdepth as second argument. But there was already
a function that used two arguments with releaseSurface as the second
argument which now was interpreted as useHighBitdepth.

This caused an issue in MediaCodecPlayerTest which used the two
argument function to pass releaseSurface.

To fix this, all the three arguments are now passed in clearSurface().

Bug: 254078351
Test: atest CtsMediaPlayerTestCases:MediaCodecPlayerTest

Change-Id: I341b9931bc4427a717f7ee750282ac4c02571c26