Check that WallpaperService callback methods are called by the main thread
We mean by 'callback methods' all the methods on the public API that start with 'on'; for example 'onCreate' or 'onSurfaceCreated'.

The new TestWallpaperService adds the layer of verification. The fake WallpaperServices used in CTS now extend this class.

A new CTS has been added to make sure that all callback methods are called by CTS.

A few other verification have been added to ensure the lifecycle
functions are called in the right order, for example:
   - many methods should only be called after 'onSurfaceCreated'
   - 'onSurfaceCreated' should only be called after 'onCreate'
   - 'onDestroy' should not be called if 'onCreate' was not called.

Bug: 240112941
Test: atest CtsWallpaperTestCases
Change-Id: I157a2c9640381979f01ce36fe913748c8626154f