Create a PublishAttributes class that tells the result of a publish request, including sip debug information
In order to more debug issues during the publication, the caller should have access to specfic SIP signaling information.
To access SIP information, using the created SipDetails class and set it as a variable in PublishAttributes class.
And the existing interface change to set the SipDetails class as a parameter.

Bug: b/238192081
Test: cts CtsTelephonyTestCases:ImsServiceTest

Change-Id: Ic99c5237724b55e8c017258be72a42226fdaa5bb