Add test cases for creating the same kind of detector at the same time.
There are two kinds of detectors, one is AlwaysOnHotwordDetector,
the other is SoftwareHotwordDetector. We don't allow to create
multiple AlwaysOnHotwordDetectors at the same time or create
multiple SoftwareHotwordDetectors at the same time.

So we need to remove this test case :
testAlwaysOnHotwordDetectorCreate_createMultipleUniqueDetectors

We also add the test cases to make sure the same kind of detector
doesn't be created twice at the same time.

Test: atest CtsVoiceInteractionTestCases
Bug: 193232191
Bug: 243598246
Change-Id: I0cc95dce2266bc0348c830fe4be72c0b84c0b70d