webkit: Use onMainThreadSync directly instead of runOnUiThread
I originally added runOnUiThread
because I had an assumption that
the Sdk Sandbox tests we are going add
would need to operate differently on the UI thread.
runOnUiThread was added so that
it could be overridden for the SDK
Sandbox environment.

This assumption was false.
I used WebkitUtils.onMainThreadSync inside
the SDK Sandbox and confirmed that this operates correctly
so switching all these calls. This sets
a precedent that will be easier to follow in
other WebView tests as we transition them
to support running in the SDK Sandbox.

Test: atest CtsWebkitTestCases:WebViewTest
Bug: 254650689
Change-Id: I0e626c5598e15b0aa1d59db6de3a290cbe770106