Temporarily ignore timing out broadcast test.
This test can easily take more than 60 seconds for the secondary user
to finish completely booting before it can hear about the broadcasts
that it's interested in.

The ideal fix would be for @EnsureHasUserAnnotation to wait for any
newly created users to settle down before it proceeds with actual
tests, since it's in a better position to wait for that one-time
event, instead of unfairly blaming the wait-for-idle or -barrier
duration on the first test that happened to be picked.

Bug: 254763137
Test: TH
Change-Id: I76e6e6c701580b1eae37fc59bf36932f3e94dfb2