Change to another active source before testing IgnoreBroadcastedFromSameSource.
If the DUT is already the active source, when pressing Home Key, it won't broadcast "Active Source" message.
We want to test if the device will send "Active Source" when pressing Home Key too, so we change the active source to another device first, to avoid the situations that DUT is already the active source.

Bug: 252963258
Test: Manually by MTK
Change-Id: I7cff2e5f93121d2581b31152d19932931c20e4c9