Ignore replies with mismatched tokens in CommandSession
We receive replies in Binder threads, so throwing there will lead to
test process crashes, while at most we just want test failures. The test
will fail if it doesn't receive the waited reply, but if the requested
activity relaunches for any reason the receiver may receive more than
one reply and also lead to this place.

In that case the test may not always fail, because it may not assert on
the lifecycle of the activity. This may not be a desirable behavior
change, but I would argue it's still better than crashing the process
and failing all tests that are yet to run. We shouldn't rely on this
crash to catch system issues anyway, because it's a very vague signal
about whether the system is actually doing something wrong.

Bug: 254712131
Test: Builds and presubmit tests.
Change-Id: I6f1721fe76db2de2dffed909f110eb7b2e1b3c2a