Change BackgroundDexOptimizationTest to reference mLastExecutionStartUptimeMs
As part of the changes for deprecating the durationIncludingSleep for
the background DexOpt metrics (as it will always be equal to the normal
duration due to the job scheduler acquiring a wake lock), we are
removing mLastExecutionStartTimeMs which is referenced in the CTS called
BackgroundDexOptimizationTest. This CL changes the test code to retrieve
mLastExecutionStartUptimeMs instead.

Bug: 255759883
Test: CTS
Merged-In: Id56926cfe9b8c51b56242ca5b25a6213bd04e25f
Change-Id: Ibab70f2e91271c67fa7c602a95d89c7d249be63a