webkit: Add shared SDK Runtime WebServer tests
Adding abstractions around the WebKit test server
so that we can invoke these tests from within the
SDK Runtime as well. Tried to keep the API as close
as possible to the regular web server so that we don't
have to adjust many tests.

Two tests fail for the SDK runtime because we can't
inject input events into the view owned by the SDK
sandbox.

Test: atest CtsSdkSandboxWebkitTestCases && atest CtsWebkitTestCases
Bug: 255745882
Change-Id: Iddabd4db0f2e9386a6ac07981ea50f3dbaa791c6