Refactor HotwordDetectionServiceNonExistenceTest
Currently, the voice cts communication across three packages. It
makes the test hard to read and we need to use broadcast, service
to let the data can pass between these thress packages.

To reduce test complexability, we will try to make the test
communication in the same test package that can access service
variables or methods easier.

The MainInteractionService doesn't define a HotwordDetectionService,
HotwordDetectionServiceNonExistenceTest is used to test the create
alwayson hotword detector behavior for this case. We only move the
logic that is related to HotwordDetectionServiceNonExistenceTest back
to CtsMainVoiceInteractionService, we will remove the out package
MainInteractionService when refacor is done.

Bug: 249287756
Test: atest AlwaysOnHotwordDetectorTest
Test: atest HotwordDetectionServiceNonExistenceTest
Change-Id: I50cc1271e6bf5e4c5dc0049b4d400674b06a59f1