Fix GSI test breakages in ContentProviderAppVisibilityTests
This test was using a hardcoded package installer name to launch an
intent. On non AOSP builds, the package name for the installer is
`com.google.android.packageinstaller`. Whereas, on an AOSP build, it is
`com.android.packageinstaller`.
Now, we dynamically fetch the installer's package name to launch intents
in this test

Bug: 244772012
Test: atest CtsContentProviderTestsWithoutVisibility CtsContentProviderTestsWithVisibility
Change-Id: I5c2c477b4b54879d0968fc53d781beba9fc70f66