Verify the CC enable is updated after removing package from allowlist
We had an old bug that indexOutOfboundException occurs when setting
allowlist, this may cause the CC enable state will not be updated
correctly. The expected is if there is a Session for the Activity,
when the service app package or Activity is removed from the
allowlist, the enabled status should be updated.

Bug: 256998476
Test: atest BlankActivityTest
BYPASS_INCLUSIVE_LANGUAGE_REASON=existing API

Change-Id: I99aad079308fd204cd7607371afbd260c82a3726