Move tests from HotwordDetectionServiceBasicTest to new test (5/5)
Currently, the voice cts communication across three packages. It
makes the test hard to read and we need to use broadcast, service
to let the data can pass between these thress packages.

To reduce test complexability, we will try to make the test
communication in the same test package that can access service
variables or methods easier.

This is the final change to move tests from to new test class
HotwordDetectionServiceTest. To make the reviewer review easier (less
diff), the test renamed back to HotwordDetectionServiceBasicTest will
be in the next change.

Bug: 249287756
Test: atest CtsVoiceInteractionTestCases
Change-Id: Ieace4c1d8057b5b53cdcac6660482bf9bc8e6de2