Add CTS tests to verify share identity getLaunchedFrom APIs
Android 14 introduces APIs to allow an app to share its identity with
a launched activity and for that activity to obtain the launching
app's uid and package name. This commit adds new tests that verify
the new getLaunchedFrom APIs only return a valid uid and package name
when the launching app has opted in to sharing its identity by
calling ActivityOptions#setShareIdentityEnabled(true) or when the
activity is launched within its own uid.

Bug: 259743961
Test: atest CtsAndroidAppTestCases:android.app.cts.ShareIdentityTest
Change-Id: If958c24d51f6ba53c4cf5116594023223624ac8c