Fix the CTS test case failed issue
AppLocalesBackupTest#testBackupRestore_allAppsInstalledNoAppLocalesSet_restoresImmediately run failed because the app locales will be cleared when the app is upgraded, causing the app not to perform the backup and restore.
The solution is adding the localeConfig in the resources of the test apk to avoid the app locales being cleared after the apk upgrades.

Bug: 262340416
Test: atest AppLocalesBackupTest
Change-Id: I0594cfe0e63862f0562cbfd548c3fa9d94624f85