Fix high fail rate AlwaysOnHotwordDetectorTest
The CTS test will update the VoiceInteractionService to CTS test
VoiceInteractionService. The original VIS tried to destroy the
detector and stopRecognition. When the test fails, we can find
the old destroy doesn't finish yet but the test tried to start a new
one. To resolve the issue, we wait for a timeout to let the previous
Hotword detection service finish all tasks then start to test.

Bug: 261677776
Test: atest AlwaysOnHotwordDetectorTest --iterations 20
Change-Id: I81165c0ff05fa4d29737ab987bbd02d50d9cb46a