Accepting alternative minWith/minHeight by current display density
The display density was set to FHD while device booting up, and
that was done even before PackageManager started. So, the activity's
minWidth/minHeight in the ActivityInfo was calculated based on
FHD resolution afterward, which is different from the one that
calculated from the device stable density in the test.

Bug: 261123586
Test: atest ManifestLayoutTests
Change-Id: Ibae62763899f10aebc330541ed65d9e59d895dee