Install ShareIdentityTest helper app in full app mode
The ShareIdentityTest makes use of a test helper app to configure
the environment for each test and call back into the test's activity
to ensure the new Share App Identity APIs work as expected. When
the test is run in instant mode, the test helper app is not available
to the test app. This commit resolves this by always ensuring the
test helper app is installed in full mode; the helper app's activity
is already visible to instant apps, and this ensures the instant
app test can invoke the helper activity to drive the test.

Fixes: 262818905
Test: atest --instant ShareIdentityTest
Change-Id: I17f493f329c19a53b2fb75424c90af1ba398c9e3