Add CTS tests for touchscreen coordinate mapping in various orientations
When the touch screen has exactly the same dimensions as the display,
each point in the raw input device's coordinates should correspond to
one pixel on the display.

Add tests to ensure that the touch screen maps precisely to the display
pixels in all of the four orientations. We do this by verifiying that
taps on the all four corners of the touch screen map to the
corresponding display pixel in all four orientations.

Bug: 257118693
Test: atest TouchScreenTest
Change-Id: Ia54072f2a0500461e6a1be86bce57440bb2eca86