Update ShareIdentityTest with queries tag for test app
Android U introduced new APIs to allow a calling app to share its
identity with a launched activity. Currently these APIs do not
support enabling app visibility; if all of the ShareIdentityTest
cases are run together and the startActivityForResult test is run
first, then app visibility for the test app is granted to the
ShareIdentityTest app. However, if the test to verify the explicit
sharing is run in isolation, then app visibility is not granted and
ShareIdentityTest cannot obtain the uid of the test app. This commit
adds a queries tag to the manifest of the ShareIdentityTest app to
allow it to query the test app; once app visibility is supported by
the share identity APIs, this tag will be removed.

Fixes: 263167263
Test: atest ShareIdentityTest#testShareIdentity_explicitIdentityShared_identityAvailableToActivity
Change-Id: Id9a882ab210a6060e86b71a129988e97fdbdab9b