Add tests to verify app visibility after share identity opt-in
Android U introduced new APIs to allow a launching app to opt-in to
sharing its identity with a launched activity; this opt-in should
include app visibility. This commit adds tests to verify an app's
details and signing identity are available to the launched activity
if the launching app has opted-in to sharing its identity.

Bug: 262605364
Test: atest ShareIdentityTest
Change-Id: I596081ce42db9c9153a47b9024e1834d2441aedc