Modified so that the third call can be stored in mCall3.
When tested in cf, as a timing issue, 'mCall3 = getCall(mCurrentCallId);' saved TelephonyConnectionService@2, not the third call.

If the URI string value of mCall3 and the third call's URI string(used for calling TelecomManager#placeCall) are different, wait until next onCallAdded and then modify it so that the third call can be properly saved.

Bug: 262969406
Test: atest "CtsTelephonyTestCases:android.telephony.ims.cts.ImsCallingTest#testCallJoinExistingConferenceCallAfterCallSwap" --iterations 100(with cf)
Test: atest "CtsTelephonyTestCases:android.telephony.ims.cts.ImsCallingTest"(with cf)
Change-Id: I3ee21468bfc068293aa3077819b5b5fc92f48ddd