Refactor HotwordDetectionServicePermissionTest
Currently, the voice cts communication across three packages. It
makes the test hard to read and we need to use broadcast, service
to let the data can pass between these thress packages.

To reduce test complexability, we make the test communication in
the same test package that can access service variables or methods
easier.

This change refactors the HotwordDetectionServicePermissionTest to
use the new infrastructure and delete the files in out service
package.

Bug: 249287756
Bug: 263687894
Test: atest CtsVoiceInteractionTestCases
Change-Id: Ia404129a18db31b934c9b4f7ff7d0dc819627030